% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_ts}
\alias{extract_ts}
\title{Extract gap filtered time series}
\usage{
extract_ts(x = cl_streamflow_mm, md, date_ranges, c_d_min = 80,
  c_m_min = 80, lg_max = 6, excelfile = NULL, outdir = NULL)
}
\arguments{
\item{x}{x daily timeseries (tibble) with Date and data columns named with the stationcode}

\item{md}{metadata table (tibble) with a stationcode column containing the stations to be used}

\item{date_ranges}{date range(s) (character or character vector) formatted as "start-end" in years}

\item{c_d_min}{minimum daily completeness}

\item{c_m_min}{minimum monthly completeness}

\item{lg_max}{maximum gap length}

\item{excelfile}{file path to write excel table with metadata per date range as workbooks, optional}

\item{outdir}{directory path for saving timeseries into univariate files (see \code{write_univariate}), optional}
}
\value{
a list of two objects:
\itemize{
\item streamflow: a list of daily time series, named by date ranges
\item metadata: a list of metadata tables, named by date ranges
}
}
\description{
The metadata table must have gap statistics columns matching the date range(s), as created by \code{gap_statistics}.
}
