% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_ts_from_md}
\alias{filter_ts_from_md}
\title{Filter times series by stationcode and/or date range}
\usage{
filter_ts_from_md(x, md = NULL, date_range = NULL)
}
\arguments{
\item{x}{daily timeseries (tibble) with Date and data columns named with the stationcode}

\item{md}{metadata table (tibble) with a stationcode column containing the stations to be extracted}

\item{date_range}{date range (character) formatted as "start-end" in years, optional}
}
\value{
x subset according to md and date_range (tibble)
}
\description{
Filter times series by stationcode and/or date range
}
