% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gap_statistics}
\alias{gap_statistics}
\title{Calculate Gap Statistics for different time ranges}
\usage{
gap_statistics(x, date_ranges, completeness_threshold = 100,
  md = cl_catchment_md)
}
\arguments{
\item{x}{daily timeseries as tibble with Date and data columns named with the stationcode}

\item{date_ranges}{character vector of date ranges}

\item{completeness_threshold}{completeness percentage below which a month is marked as NA when calculating monthly statistics (c_m, l_g)}

\item{md}{metadata table (tibble) with stationcode column}
}
\value{
the metadata table \code{md} with the following vars added: stationcode + c_d, c_m, l_g for each date range
}
\description{
Given a univariate daily time series, the following statistics are computed for the specified date ranges:
\itemize{
\item Completeness Percentage, daily (c_d): Percentage of non-NA values in the time series
\item Completeness Percentage, monthly (c_m): Percentage of non-NA values after aggregating to monthly
\item Longest Gap,monthly (l_g): Longest consecutive stretch of NA values after aggregating to monthly
}
}
\details{
The time ranges are specified in years as "start-end", with "2000-2009" meaning from 2000-01-01 to 2009-12-31.
}
