% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_complete}
\alias{is_complete}
\title{Assert completeness by threshold in percent of of non-NA values}
\usage{
is_complete(x, na.rm)
}
\arguments{
\item{x}{vector of values}

\item{na.rm}{completeness percentage below which the input should be considered NA}
}
\value{
TRUE if less than na.rm percent of values are NA, FALSE otherwise
}
\description{
Assert completeness by threshold in percent of of non-NA values
}
