% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longest_gap}
\alias{longest_gap}
\title{Calculate Longest Contiguous Stretch of NAs}
\usage{
longest_gap(x, logical = FALSE)
}
\arguments{
\item{x}{vector of values}

\item{logical}{if TRUE, x must be a logical vector with FALSE representing NA}
}
\value{
length of longest contiguous Stretch of NAs
}
\description{
Calculate Longest Contiguous Stretch of NAs
}
