% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent_complete}
\alias{percent_complete}
\title{Calculate completeness as percentage of non-NA values}
\usage{
percent_complete(x, logical = FALSE)
}
\arguments{
\item{x}{vector of values}

\item{logical}{if TRUE, x must be a logical vector with FALSE representing NA}
}
\value{
completeness in percent rounded to one decimal place
}
\description{
Calculate completeness as percentage of non-NA values
}
