% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_univariate}
\alias{write_univariate}
\title{Write out multivariate time series into separate files per variable}
\usage{
write_univariate(x, outdir, md = NULL, date_range = NULL)
}
\arguments{
\item{x}{daily timeseries (tibble) with Date and data columns named with the stationcode}

\item{outdir}{directory path to write}

\item{md}{metadata table (tibble) with a stationcode column containing the stations to be used, optional}

\item{date_range}{date range (character) formatted as "start-end" in years, optional}
}
\value{
invisible, writes variables of \code{x} into seperate files named <stationcode>.csv (delimited by space) into \code{outdir}.
}
\description{
If a metadata table is given, only the stationcodes present in the stationcode column will be written out.
If a date range is given, the time series is first subset accordingly (e.g. "2000-2004" translates to 2000-01-01-2004-12-31),
and files will be written into a sub-directory of \code{outdir} named by \code{date_range}.
}
